move unused code
