+++ deprecated/imageGenerator.ts
