+++ deprecated/sas.ts
