+++ deprecated/schema.ts
